+++ semgrep/semgrep/core_output.py
@@ -365,7 +365,6 @@ class CoreOutput:
                 start=match.start,
                 end=match.end,
                 extra=match.extra,
-                lines_cache={},
             )
             return rule_match